modified stock variable to parseFloat instead of parseInt, because items with float quantities would not pass the stock condition on below 1 quantities

+++ src/adminComponents/SalesControl.jsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, {Component} from 'react';
 
 import CustomSelect from '../globalComponents/CustomSelect';
 import $ from 'jquery';
@@ -80,7 +80,7 @@ class SalesControl extends Component {
             })
                 .then(res => res.json())
                 .then(res => {
-                    let stock = parseInt(res.data[0].stock);
+                    let stock = parseFloat(res.data[0].stock);
                     if (stock > 0 && stock >= this.state.quantity) {
                         let product = this.state.selectedProduct;
 
@@ -99,11 +99,11 @@ class SalesControl extends Component {
                         product.totalBs = this.roundToNiceNumber(product.priceBs * product.quantity - product.priceBs * (discount / 100));
 
                         if (discount) {
-                            product.totalFormattedDollars = <span className={discount ? 'text-danger' : ''}>{Intl.NumberFormat('es-VE', { style: 'currency', currency: 'USD' }).format(product.totalDollars) + ` (-${discount}%)`}</span>;
-                            product.totalFormattedBs = <span className={discount ? 'text-danger' : ''}>{Intl.NumberFormat('es-VE', { style: 'currency', currency: 'VES' }).format(product.totalBs) + ` (-${discount}%)`}</span>;
+                            product.totalFormattedDollars = <span className={discount ? 'text-danger' : ''}>{Intl.NumberFormat('es-VE', {style: 'currency', currency: 'USD'}).format(product.totalDollars) + ` (-${discount}%)`}</span>;
+                            product.totalFormattedBs = <span className={discount ? 'text-danger' : ''}>{Intl.NumberFormat('es-VE', {style: 'currency', currency: 'VES'}).format(product.totalBs) + ` (-${discount}%)`}</span>;
                         } else {
-                            product.totalFormattedDollars = Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD' }).format(product.totalDollars);
-                            product.totalFormattedBs = Intl.NumberFormat('es-VE', { style: 'currency', currency: 'VES' }).format(product.totalBs);
+                            product.totalFormattedDollars = Intl.NumberFormat('en-US', {style: 'currency', currency: 'USD'}).format(product.totalDollars);
+                            product.totalFormattedBs = Intl.NumberFormat('es-VE', {style: 'currency', currency: 'VES'}).format(product.totalBs);
                         }
 
                         productFormatted = {
@@ -161,7 +161,7 @@ class SalesControl extends Component {
                             var index = this.state.addedProducts.findIndex(product => product.id === productFormatted.id);
                             this.setState(
                                 {
-                                    addedProducts: [...this.state.addedProducts.slice(0, index), Object.assign({}, this.state.addedProducts[index], { ...productFormatted }), ...this.state.addedProducts.slice(index + 1)],
+                                    addedProducts: [...this.state.addedProducts.slice(0, index), Object.assign({}, this.state.addedProducts[index], {...productFormatted}), ...this.state.addedProducts.slice(index + 1)],
                                     quantity: 1,
                                     selectedProduct: null,
                                     stockError: '',
@@ -362,10 +362,10 @@ class SalesControl extends Component {
                                         <th scope="col">Precio Bs</th>
                                         <th scope="col">Cantidad</th>
                                         <th scope="col">
-                                            Total<span className="font-weight-bold text-warning">{' ' + Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD' }).format(this.state.totalDollars)}</span>
+                                            Total<span className="font-weight-bold text-warning">{' ' + Intl.NumberFormat('en-US', {style: 'currency', currency: 'USD'}).format(this.state.totalDollars)}</span>
                                         </th>
                                         <th scope="col">
-                                            Total<span className="font-weight-bold text-warning">{' ' + Intl.NumberFormat('es-VE', { style: 'currency', currency: 'VES' }).format(this.state.totalBs)}</span>
+                                            Total<span className="font-weight-bold text-warning">{' ' + Intl.NumberFormat('es-VE', {style: 'currency', currency: 'VES'}).format(this.state.totalBs)}</span>
                                         </th>
                                     </tr>
                                 </thead>
@@ -376,11 +376,11 @@ class SalesControl extends Component {
                                                 <th>{index + 1}</th>
                                                 <th>{product.id}</th>
                                                 <th>
-                                                    <img className="" style={{ maxWidth: '40px' }} src={product.imagePath} />
+                                                    <img className="" style={{maxWidth: '40px'}} src={product.imagePath} />
                                                     {product.name}
                                                 </th>
-                                                <th>{Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD' }).format(product.price)}</th>
-                                                <th>{Intl.NumberFormat('es-VE', { style: 'currency', currency: 'VES' }).format(product.priceBs)}</th>
+                                                <th>{Intl.NumberFormat('en-US', {style: 'currency', currency: 'USD'}).format(product.price)}</th>
+                                                <th>{Intl.NumberFormat('es-VE', {style: 'currency', currency: 'VES'}).format(product.priceBs)}</th>
                                                 <th>{product.quantity}</th>
                                                 <th>{product.totalFormattedDollars}</th>
                                                 <th>{product.totalFormattedBs}</th>
